Type:feat 添加适配更多UT 子项目
Description:添加适配更多UT 子项目

RootCause:Null

Solution: 在脚本中添加更多文管子项目适配

Former-commit-id: 64c107573cc72ed7a0e3f7e2f10a7223f396554a

+++ tests/test-prj-running.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# 获取参数
+CLEAR_BUILD="clear" #是否清场 not_clear 就不清场
+UT_PRJ_TYPE=${1} #运行UT类型 ut_all dde_file_manager dde-file-manager-lib dde-desktop dde-dock-plugins dde-file-manager-plugins dde-file-thumbnail-tool deepin-anything-server-plugins
+
 # 定位脚本所在当前目录
 TESTS_FOLDER=$(cd "$(dirname "$0")";pwd)
 echo $TESTS_FOLDER
@@ -13,7 +17,10 @@ BUILD_DIR=$PROJECT_FOLDER/build-ut
 echo $BUILD_DIR
 
 # 清场,删除编译文件夹
-rm -rf $BUILD_DIR
+if [ "$CLEAR_BUILD" != "not_clear" ] ; then
+    rm -rf $BUILD_DIR
+    echo "clear " $BUILD_DIR
+fi
 # 创建文件夹
 mkdir -p $BUILD_DIR
 
@@ -22,7 +29,12 @@ echo `pwd`
 echo "start dde-file-manager project all CI cases"
 
 # 下面是UT相关的case
-
-./all-ut-prj-running.sh $PROJECT_FOLDER $BUILD_DIR
+if [ ! -n "$UT_PRJ_TYPE" ] ; then
+    echo "ut case type is all"
+    ./all-ut-prj-running.sh $PROJECT_FOLDER $BUILD_DIR "ut_all"
+else
+    echo "ut case type is " $UT_PRJ_TYPE
+    ./all-ut-prj-running.sh $PROJECT_FOLDER $BUILD_DIR $UT_PRJ_TYPE
+fi
 
 exit 0
